refactor: rename env to app context
