+++ lib/contexts/app.js
@@ -6,10 +6,10 @@ import { createRouter } from "@lib/routers";
  * 실행 환경을 위한 컨텍스트
  * @description 현재 라우터 타입, 라우터, 유저 인증 서비스를 관리
  */
-class RuntimeEnviroment {
+class AppContext {
   constructor() {
-    if (RuntimeEnviroment.instance) return RuntimeEnviroment.instance;
-    RuntimeEnviroment.instance = this;
+    if (AppContext.instance) return AppContext.instance;
+    AppContext.instance = this;
 
     this.router = null;
     this.routerType = null;
@@ -23,5 +23,5 @@ class RuntimeEnviroment {
   }
 }
 
-const env = new RuntimeEnviroment();
-export default env;
+const app = new AppContext();
+export default app;
