+++ src/pages/(app)/layout.js
@@ -1,4 +1,4 @@
-import env from "@lib/contexts/env";
+import env from "@lib/contexts/app";
 import { Header, Nav, Footer } from "@components";
 
 export default function Layout(container, { children }) {
